Update index-responsive.css
explanation for previous commit (I acidentaly clicked commit)
for some reason the image tag in the css grid element does not find its space limits from its parents and when I set it to 100% height it takes up way more space then it should bassed on the space that it was given by its parent element which is a css grid div, what ends up happening is that image desides its size and the grid adapts around it when it should be the other way round.
I will have to find a way to replace the image tags with background images on the divs with rounded corners something that when I tried previously would not render when I set the background image.