the new css bits are now fully responsive
I added a few new break points to make it better  though most are empty and more for scafolding.
I also chandged the text for the title to be a bit bigger, it was a bit to small.
I als capitaliszed Testing in the title section
there were also 2 instances of <meta charset=UTF-8> so I removed 1
from what I can tell this is ready to be implamented in the main pages, so that is what im doing next.