went to update the index and start implamenting the new css but ran into issues
I did not have the link section so I have to implament that and have run into some weird issues so im commiting what I have right. when I eneable the class that centers the text in the midle of the description area  it freaks out and takes up way to much space.
also added a formating rule to the readme file