worked some more on the styling for the forms
added borders to the feild sets, updated the colors and added styling for the legend though that is no were close to being done, I want to center it but I cant find a good way to do so without using flexbox which would introduce to many divs.
I also added basic styling for label and input type button along with adding input type month to the already existant list which I have also made to have no border.