added more form elements to the css
I added more form elements to the css and ran int an issue were option elements inside of datalists do not take styling from the option style
also, in a few of the css styles I used background to set the background color. it workded but I wanted to have them all be more spacific, so now they are.
again non of the styles are final