implamented a switch for dark an light theme, it is ugly, will fix in the future
added code necessary for switching themes
had to put the content of the page in a div and am using input type checkbox for the switcher, this might change. the looks are not final for the themes them selves and the switch button.
it is CSS only and a bit convoluted, if I wind a different, also CSS only method that is less convoluted I will.