added hoverable side navigation menu for nav menu and theme switcher
added hoverable side navigation menu for nav menu and theme switcher
this took me a while because I had to settle for the least preferable method to achieve this, I added a file to the repo in the docs directory.
the file is CSS grid implementation of nav bar.md
the `input type="checkbox"` is still visible though I plant to fix that next.
the CSS is also inefficient I and fix that soon
non of these styles are final.