added a new effect to links, it is not 100% done yet
added a new effect to links, it is not 100% done yet
links have a small underlined bar under them that when you hover expands to 100% of the size of the text of the link.
this is not 100% completed yet, the basics are there but I need to  make it such that the link does effect does not work with link images and find a better way to exclude the effect from the nav overlay.
non of these styles are final.